Setting up files and creating the display
I created a basic HTML file to contain and drive our code. It imports
index.css and index.js, and for now index.css does all of the real work.

The only thing notable about it is that it has a div#tetris that we will
use as a container for the game, which our CSS can style.

index.js is empty right now, it was only added so that it won't need to
be added later.

index.css introduces a handful of CSS variables that I'll use to keep
the code sane. But mostly, right now what it does is define the size and
shape of the game area - it should take up 90% of the height of the
screen, and be sized at 20 blocks tall by 10 blocks wide. A block, then,
is going to be 1/20th of the height of the screen, and the width of the
display is calculated from the size of the block, and the number of
blocks wide that the display should be.

I also made the background blue, just so that I can see it, since it has
no content right now. I could have added a border to it and accomplished
the same thing.

Next, I'll create a single block (each Tetris piece is made up of four
blocks), and place it in the display using Javascript.