Creating a block
Most of the work here is in index.js. I get the #tetris container as a
variable so that I can add blocks to it as children later.

More importantly, I create some functions that make elements for blocks,
and manipulate their locations. Blocks can be added to the display via
appendChild on the #tetris container.

Then I test them out with the code at the bottom of index.js

In index.css, I create some styling that will position each block based
on the --x-coord and --y-coord CSS variables, which are controlled in
index.js.

Next, I'll add a basic game loop that will make blocks descend.